% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{benthclim}
\alias{benthclim}
\title{Example climate at occurrence locations for Eucalyptus benthamii}
\format{An object of class \code{climvardf} (inherits from \code{data.frame}) with 13 rows and 28 columns.}
\usage{
benthclim
}
\description{
Example climate at occurrence locations for Eucalyptus benthamii
}
\author{
Remko Duursma
}
\keyword{datasets}
