% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{benthocc}
\alias{benthocc}
\title{Example species occurrences for Eucalyptus benthamii}
\format{An object of class \code{data.frame} with 779 rows and 3 columns.}
\usage{
benthocc
}
\description{
Example species occurrences for Eucalyptus benthamii
}
\author{
Remko Duursma
}
\keyword{datasets}
