% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{benthoccras}
\alias{benthoccras}
\title{Example rasterized species occurrences for Eucalyptus benthamii}
\format{An object of class \code{data.frame} with 13 rows and 3 columns.}
\usage{
benthoccras
}
\description{
Example rasterized species occurrences for Eucalyptus benthamii
}
\author{
Remko Duursma
}
\keyword{datasets}
