% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldclim.R
\name{climate_presence}
\alias{climate_presence}
\title{Long-term climate at species occurrences}
\usage{
climate_presence(species, database = c("ALA", "GBIF", "both"), vars = NULL,
  output = c("monthly", "annual"), rasterize = TRUE, ala_args = NULL,
  gbif_args = NULL)
}
\arguments{
\item{species}{Latin binomial for species, can be a vector.}

\item{database}{Either 'ALA' or 'GBIF'}

\item{vars}{Vector of climate variables to be extracted; see Details.}

\item{output}{Either 'monthly' (default), returning monthly climate data, or 'annual', in which case \code{\link{annualize_clim}} will be run on the result.}

\item{rasterize}{If TRUE, the default, resamples species occurrences into same spatial scale as Worldclim data.}

\item{ala_args}{Further arguments passed to \code{occurrences} from the \code{ALA4R} package (named list).}

\item{gbif_args}{Further arguments passed to \code{occ_search} from the \code{rgbif} package (named list).}
}
\description{
Simple interface to download species occurrences, and extract climate observations from Worldclim for where the species is present.
}
\details{
The climate variables are any of the Worldclim 2.0 variables, including "tmin","tmax","tavg","prec","srad","wind","vapr". Also allowed is 'pet', which will be extracted from the Zomer PET database (see \code{\link{get_zomer_pet}}).
}
\author{
Remko Duursma
}
