% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldclim.R
\name{get_climate_vars}
\alias{get_climate_vars}
\alias{aggregate_clim}
\alias{annualize_clim}
\title{Get worldclim observations for specific locations}
\usage{
get_climate_vars(data, vars = NULL, output = c("monthly", "annual"))

aggregate_clim(cl, funs = c("mean", "quantile"), probs = c(0.05, 0.95))

annualize_clim(cl)
}
\arguments{
\item{data}{Dataframe with columns \code{latitude} and \code{longitude}.}

\item{vars}{Vector of climate variables to be extracted; see Details.}

\item{output}{Either 'monthly' (default), or 'annual', in which case the \code{annualize} function is executed on the result. Alternatively you can generate monthly output, and use \code{annualize_clim} yourself (see Examples).}

\item{cl}{Output of \code{get_climate_vars}.}

\item{funs}{For \code{aggregate_clim}, a character vector of functions to be applied to each column.}

\item{probs}{For \code{aggregate_clim}, if 'quantile' is one of the functions, the probabilities to be calculated (see \code{\link{quantile}})}
}
\description{
Given a dataframe with longitude and latitude coordinates, extracts rainfall and temperature data from Worldclim. Rasters are downloaded if not provided.
}
\details{
The climate variables are any of the Worldclim 2.0 variables, including "tmin","tmax","tavg","prec","srad","wind","vapr". Also allowed is 'pet', which will be extracted from the Zomer PET database (see \code{\link{get_zomer_pet}}).
}
\examples{
\dontrun{
# First set the path to where WorldClim layers will be downloaded,
# and if extracting PET, where you downloaded the Zomer PET database
# see ?get_zomer_pet
options(worldclimpath="/path/to/worldclim")
res <- get_climate_vars(data=data.frame(longitude=150, latitude=-33), vars=c("tmin","bio"))

# To extract PET, first set the path to where the layers were downloaded
options(zomerpetpath="c:/data/zomer")
res <- get_climate_vars(data=data.frame(longitude=150, latitude=-33), vars=c("pet","tavg"))

# By default all monthly values are returned, use annualize to get annual
# averages (or total in case of MAP):
annualize(res)

# ... or simply use output="annual"
res <- get_climate_vars(data=data.frame(longitude=150, latitude=-33),
vars=c("pet","tavg"), output="annual")
}
}
\author{
Remko Duursma
}
