% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurrences.R
\name{get_occurrences_ala}
\alias{get_occurrences_ala}
\alias{get_occurrences_gbif}
\title{Get species occurrences from the ALA or GBIF}
\usage{
get_occurrences_ala(species, ala_args = NULL)

get_occurrences_gbif(species, gbif_args = NULL)
}
\arguments{
\item{species}{The full scientific species name, quoted.}

\item{ala_args}{Optionally, a list of arguments passed to \code{occurrences} (ALA4R package).}

\item{gbif_args}{Optionally, a list of arguments passed to \code{occ_search} (rgbif package).}
}
\description{
A simple wrapper to \code{\link{occurrences}} from \CRANpkg{ALA4R}. Performs a search for a \code{species} (e.g. 'Eucalyptus globulus'), sets the \code{download_reason} to 'testing' (see \code{?ala_reasons}), and returns only the species, the longitude, and the latitude of all observations.
}
\details{
Note that when passing additional arguments to \code{occ_search}, the following arguments are already set and can not be overriden : \code{hasGeospatialIssue} (FALSE), \code{limit} (50000), \code{fields} (species, lat and long), and of course \code{scientificName}. When passing additional arguments to \code{occurrences}, only \code{download_reason} (7) is already set.
}
\examples{
\dontrun{
o <- get_occurrences_ala("Eucalyptus botryoides")

# Should return a message like:
# INFO [2017-05-12 09:52:10] ALA returned 3274 records for Eucalyptus botryoides in 2.6 sec.
}
}
\seealso{
\code{\link{get_occurrences_gbif}}
}
\author{
Remko Duursma
}
