% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurrences.R
\name{get_occurrences_gbif}
\alias{get_occurrences_gbif}
\title{Obtain species occurrences from GBIF}
\usage{
get_occurrences_gbif(species)
}
\description{
Very simple wrapper to \code{\link{occ_search}}
}
\author{
Remko Duursma
}

