% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldclim.R
\name{get_worldclim_prectemp}
\alias{get_worldclim_prectemp}
\title{Get worldclim observations for specific locations}
\usage{
get_worldclim_prectemp(data, topath = tempdir(), return = c("all",
  "summary"), worldclim = NULL)
}
\arguments{
\item{data}{Dataframe with columns \code{latitude} and \code{longitude}.}

\item{topath}{Path where Worldclim rasters will be downloaded to (defaults to a temporary directory)}

\item{return}{If 'all' returns monthly tmean, precip, as well as MAT and MAP. If 'summary', returns quantiles and means of MAT and MAP across all locations in \code{data}.}

\item{worldclim}{Optionally, list of rasters returned by \code{\link{get_worldclim_rasters}}, to save time.}
}
\description{
Given a dataframe with longitude and latitude coordinates, extracts rainfall and temperature data from Worldclim. Rasters are downloaded if not provided.
}
\author{
Remko Duursma
}
