% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldclim.R
\name{get_worldclim_rasters}
\alias{get_worldclim_rasters}
\title{Download Worldclim rasters}
\usage{
get_worldclim_rasters(topath, clean = FALSE)
}
\arguments{
\item{topath}{Local path to store files. Can use \code{\link{tempdir()}} if files need not be kept.}

\item{clean}{Delete files after downloading (unimplemented)}
}
\description{
Downloads 10min resolution Worldlclim rasters, stores them for reuse, and reads them into raster objects. Currently only downloads only tmean (monthly mean temperature) and prec (monthly total precipitation) layers.
}
\details{
Zip files are downloaded from \url{http://biogeo.ucdavis.edu/data/climate/worldclim/1_4/grid/cur}
}
\examples{
# Return worldclim tmean and prec rasters, as a list with those components
\dontrun{
wc <- get_worldclim_rasters(topath=tempdir())
}
}
\author{
Remko Duursma
}
