% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zomer_pet.R
\name{get_zomer_pet}
\alias{get_zomer_pet}
\title{Extract PET from the CGIAR-CSI database}
\usage{
get_zomer_pet(data)
}
\arguments{
\item{data}{Dataframe with longitude, latitude}
}
\description{
To allow extraction of PET from the CGIAR-CSI Global aridity and PET database (\url{http://www.cgiar-csi.org/data/global-aridity-and-pet-database}), you must manually download the file 'Global PET - Annual.zip' from the above url (follow 'Download from the HarvestChoice Dropbox'), and unzip it in some directory. Then set the path with options() as shown below in the examples.
}
\examples{
\dontrun{
options(zomerpetpath="path/to/zomer")
dat <- data.frame(longitude=c(150,145,140), latitude=c(-33,-33,-33))
get_zomer_pet(dat)
}
}
