% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zomer_pet.R
\name{set_zomerpet_path}
\alias{set_zomerpet_path}
\title{Set the path where Zomer's PET database is downloaded}
\usage{
set_zomerpet_path(path)
}
\arguments{
\item{path}{Path to downloaded, unzipped, annual PET database.}
}
\description{
To allow extraction of PET from the CGIAR-CSI Global aridity and PET database (\url{http://www.cgiar-csi.org/data/global-aridity-and-pet-database}), you must manually download the file 'Global PET - Annual.zip' from the above url (follow 'Download from the HarvestChoice Dropbox'), and unzip it in some directory. When intending to use the database, set the path to the location with this function.
}
\examples{
# Directory that contains unzipped annual PET data
#(this directory will contain subdirectory 'PET_he_annual')
\dontrun{
set_zomerpath("c:/data/zomerpet")
}
}
