% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldclim.R
\name{worldclim_presence}
\alias{worldclim_presence}
\title{Long-term climate at species occurrences}
\usage{
worldclim_presence(species, database = c("ALA", "GBIF"), rasterize = TRUE,
  topath = tempdir(), return = c("summary", "all"))
}
\arguments{
\item{species}{Latin binomial for species, can be a vector.}

\item{database}{Either 'ALA' or 'GBIF'}

\item{rasterize}{If TRUE, the default, resamples species occurrences into same spatial scale as Worldclim data.}

\item{topath}{Path to store Worldclim rasters (if present there, will not be re-downloaded)}

\item{return}{If summary (the default), returns summary variables of precip and temp, one row per species. For 'all' returns climate variables (and monthly ones) for all locations; result is a list if species is a vector.}
}
\description{
Simple interface to download species occurrences, and extract climate observations from Worldclim for where the species is present.
}
\author{
Remko Duursma
}

