% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldclim.R
\name{climate_presence}
\alias{climate_presence}
\title{Long-term climate at species occurrences}
\usage{
climate_presence(species, database = c("ALA", "GBIF", "both"), vars = NULL,
  output = c("monthly", "annual"), rasterize = TRUE, ala_args = NULL,
  gbif_args = NULL)
}
\arguments{
\item{species}{Latin binomial for species, can be a vector.}

\item{database}{Either 'ALA' or 'GBIF'}

\item{vars}{Vector of climate variables to be extracted; see Details.}

\item{rasterize}{If TRUE, the default, resamples species occurrences into same spatial scale as Worldclim data.}

\item{return}{If summary (the default), returns summary variables of precip and temp, one row per species. For 'all' returns climate variables (and monthly ones) for all locations; result is a list if species is a vector.}
}
\description{
Simple interface to download species occurrences, and extract climate observations from Worldclim for where the species is present.
}
\details{
The climate variables are any of the Worldclim 2.0 variables, including "tmin","tmax","tavg","prec","srad","wind","vapr". Also allowed is 'pet', which will be extracted from the Zomer PET database (see \code{\link{get_zomer_pet}}).
}
\author{
Remko Duursma
}
