% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurrences.R
\name{get_occurrences_ala}
\alias{get_occurrences_ala}
\alias{get_occurrences_gbif}
\title{Get species occurrences from the ALA or GBIF}
\usage{
get_occurrences_ala(species)

get_occurrences_gbif(species)
}
\arguments{
\item{species}{The full scientific species name, quoted.}
}
\description{
A simple wrapper to \code{\link{occurrences}} from \CRANpkg{ALA4R}. Performs a search for a \code{species} (e.g. 'Eucalyptus globulus'), sets the \code{download_reason} to 'testing' (see \code{?ala_reasons}), and returns only the species, the longitude, and the latitude of all observations.
}
\examples{
\dontrun{
o <- get_occurrences_ala("Eucalyptus botryoides")

# Should return a message like:
# INFO [2017-05-12 09:52:10] ALA returned 3274 records for Eucalyptus botryoides in 2.6 sec.
}
}
\seealso{
\code{\link{get_occurrences_gbif}}
}
\author{
Remko Duursma
}
