% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldclim.R
\name{get_worldclim_rasters}
\alias{get_worldclim_rasters}
\title{Download Worldclim rasters}
\usage{
get_worldclim_rasters(wc_vars = "tavg")
}
\description{
Downloads and unzips 10min resolution Worldlclim rasters, stores them for reuse.
}
\details{
Zip files are downloaded from \url{http://biogeo.ucdavis.edu/data/worldclim/v2.0/tif/base}
}
\examples{
# Return worldclim tmean and prec rasters, as a list with those components
\dontrun{
wc <- get_worldclim_rasters(wc_vars=c("tavg","tmin"))
}
}
\author{
Remko Duursma
}
