% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zomer_pet.R
\name{get_zomer_pet}
\alias{get_zomer_pet}
\title{Extract PET from the CGIAR-CSI database}
\usage{
get_zomer_pet(data)
}
\description{
Extract PET from the CGIAR-CSI database
}
\examples{
dat <- data.frame(longitude=c(150,145,140), latitude=c(-33,-33,-33))
get_zomer_pet(dat)
}
